new or modified files for Dieharder 3.31.0

+++ src/list_rand.c
@@ -25,10 +25,10 @@ void list_rand()
  
  printf("#==================================================================\n");
  printf("# generator type: %s\n", gsl_rng_name(rng));
- printf("# seed value: %u, max value = %u  count = %d\n",seed, random_max,tsamples);
+ printf("# seed value: %lu, max value = %u  count = %d\n",seed, random_max, (int)tsamples);
  printf("# Count\t int rand\tuniform rand\n");
  printf("# ==================================================================\n");
- for(i = 1;i<=tsamples;i++) printf("%d\t%u\t %10.8f\n",i,gsl_rng_get(rng),gsl_rng_uniform(rng));
+ for(i = 1;i<=tsamples;i++) printf("%d\t%lu\t %10.8f\n",i,gsl_rng_get(rng),gsl_rng_uniform(rng));
 
 }
